+++ aldrin/src/low_level.rs
@@ -1,3 +1,7 @@
 //! Low-level types
 //!
 //! The types in this module are primarily intended for use by the code generator.
+
+mod event;
+
+pub use event::Event;